+++ pyaerocom/io/read_aasetal.py
@@ -8,7 +8,7 @@ import pandas as pd
 from pyaerocom.helpers import get_tot_number_of_seconds
 from pyaerocom.io.readungriddedbase import ReadUngriddedBase
 from pyaerocom.stationdata import StationData
-from pyaerocom.ungriddeddata import UngriddedData
+from pyaerocom.ungriddeddata_meta import UngriddedData
 from pyaerocom.units_helpers import convert_unit
 
 logger = logging.getLogger(__name__)
