+++ pyaerocom/ungriddeddata_meta.py
