+++ tests/io/test_read_aasetal.py
@@ -13,7 +13,7 @@ from numpy.testing import assert_almost_equal
 from pyaerocom import const
 from pyaerocom.exceptions import UnitConversionError
 from pyaerocom.io.read_aasetal import ReadAasEtal
-from pyaerocom.ungriddeddata import UngriddedData
+from pyaerocom.ungriddeddata_meta import UngriddedData
 from pyaerocom.units_helpers import convert_unit
 from tests.conftest import lustre_unavail
